golden: debian 11 efi-based packer implementation
Significant cleanup and verification of every preseed configuration

This commit within the larger PR involved verifying every preseed configuration
option using the documentation as a guide. The documentation is also referenced
now within this code.

This commit may not be precisely correct as I changed some ssh configuration. The
real goal here though is to use a file provisioner and a shell provisioner to
insert an overlay tar archive into the virtual machine and then untar it within
the broader system.

Co-Authored-by: Robert Starmer <robert@cmoputelify.com>